# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreatePostDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "default": 1, "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "type": "number" }
          },
          { "name": "courseId", "required": false, "in": "query", "schema": { "type": "string" } },
          {
            "name": "departmentId",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "yearLevel",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "maximum": 6, "type": "number" }
          },
          {
            "name": "moduleNumber",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "maximum": 3, "type": "number" }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/PostType" }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": { "$ref": "#/components/schemas/PostStatus" }
          },
          { "name": "authorId", "required": false, "in": "query", "schema": { "type": "string" } },
          {
            "name": "tagSlug",
            "required": false,
            "in": "query",
            "description": "Filter by tag slug (exact match)",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedPostEntity" }
              }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/saved": {
      "get": {
        "operationId": "PostsController_getSavedPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "default": 1, "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "minimum": 1, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedPostEntity" }
              }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/s/{shortCode}": {
      "get": {
        "operationId": "PostsController_findByShortCode",
        "parameters": [
          { "name": "shortCode", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/search": {
      "get": {
        "operationId": "PostsController_search",
        "parameters": [
          { "name": "q", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "page", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Search results with pagination" } },
        "tags": ["posts"]
      }
    },
    "/api/posts/trending": {
      "get": {
        "operationId": "PostsController_getTrendingPosts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Results per page",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "posts": [
                        {
                          "id": "post123",
                          "title": "Linear Algebra Notes",
                          "trendingScore": 42.5,
                          "views": 150,
                          "createdAt": "2026-03-19T12:00:00Z"
                        }
                      ],
                      "total": 250,
                      "page": 1,
                      "limit": 20
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostsController_findOne",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      },
      "patch": {
        "operationId": "PostsController_update",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdatePostDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/save": {
      "post": {
        "operationId": "PostsController_savePost",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "201": { "description": "" } },
        "tags": ["posts"]
      },
      "delete": {
        "operationId": "PostsController_unsavePost",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/react": {
      "post": {
        "operationId": "PostsController_react",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ReactToPostDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/status": {
      "patch": {
        "operationId": "PostsController_updateStatus",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdatePostStatusDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/summarize": {
      "post": {
        "operationId": "PostsController_summarize",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/tags": {
      "post": {
        "operationId": "PostsController_addTags",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TagPostDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PostDetailEntity" } }
            }
          }
        },
        "tags": ["posts"]
      }
    },
    "/api/posts/{id}/tags/{tagId}": {
      "delete": {
        "operationId": "PostsController_removeTag",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "tagId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["posts"]
      }
    },
...
    "/api/reading-lists/{id}/posts/{postId}": {
      "post": {
        "operationId": "ReadingListsController_addPost",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "postId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ReadingListEntity" } }
            }
          }
        },
        "tags": ["reading-lists"]
      },
      "delete": {
        "operationId": "ReadingListsController_removePost",
        "parameters": [
          { "name": "id", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "postId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["reading-lists"]
      }
    }
  },
  "info": {
    "title": "Unishare API",
    "description": "API for the Unishare platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["NOTE", "OLD_QUESTION", "ASSIGNMENT"] },
          "courseId": {
            "type": "string",
            "description": "Course ID. Must belong to the authenticated user department."
          },
          "title": { "type": "string", "minLength": 3, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 2000 },
          "isAnonymous": { "type": "boolean", "default": false },
          "externalUrl": { "type": "string", "maxLength": 500 },
          "examYear": { "type": "number", "minimum": 1900, "maximum": 2100 },
          "moduleNumber": { "type": "number", "minimum": 1, "maximum": 20 },
          "year": { "type": "number", "minimum": 1, "maximum": 6 },
          "semester": { "type": "number", "minimum": 1, "maximum": 3 },
          "tags": {
            "example": ["Linear Algebra", "Calculus"],
            "description": "Array of tag names (max 5 tags)",
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["type", "courseId", "title", "description", "moduleNumber", "year", "semester"]
      },
      "PostType": { "type": "string", "enum": ["NOTE", "OLD_QUESTION", "ASSIGNMENT"] },
      "PostStatus": { "type": "string", "enum": ["PENDING", "APPROVED", "REJECTED"] },
      "PostPublicationStatus": {
        "type": "string",
        "enum": ["PUBLISHED", "PENDING_REVIEW", "REJECTED"]
      },
      "PostAuthorDeptEntity": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
        "required": ["id", "name"]
      },
      "PostAuthorEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "image": { "type": "string", "nullable": true },
          "enrollmentYear": { "type": "number", "nullable": true },
          "department": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PostAuthorDeptEntity" }]
          }
        },
        "required": ["id", "name"]
      },
      "PostCourseDeptEntity": {
        "type": "object",
        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
        "required": ["id", "name"]
      },
      "PostCourseEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "code": { "type": "string" },
          "name": { "type": "string" },
          "department": { "$ref": "#/components/schemas/PostCourseDeptEntity" }
        },
        "required": ["id", "code", "name", "department"]
      },
      "PostFileEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "key": { "type": "string" },
          "name": { "type": "string" },
          "size": { "type": "number" },
          "mimeType": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "downloads": { "type": "number" }
        },
        "required": ["id", "key", "name", "size", "mimeType", "createdAt", "downloads"]
      },
      "PostCountEntity": {
        "type": "object",
        "properties": { "comments": { "type": "number" }, "savedBy": { "type": "number" } },
        "required": ["comments", "savedBy"]
      },
      "PostTagEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "color": { "type": "string" }
        },
        "required": ["id", "name", "slug", "color"]
      },
      "PostDetailEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "allOf": [{ "$ref": "#/components/schemas/PostType" }] },
          "status": { "allOf": [{ "$ref": "#/components/schemas/PostStatus" }] },
          "publicationStatus": {
            "allOf": [{ "$ref": "#/components/schemas/PostPublicationStatus" }]
          },
          "trendingScore": { "type": "number" },
          "isAnonymous": { "type": "boolean", "default": false },
          "title": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "externalUrl": { "type": "string", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "summarizedAt": { "format": "date-time", "type": "string", "nullable": true },
          "contentWarning": { "type": "string", "nullable": true },
          "examYear": { "type": "number", "nullable": true },
          "moduleNumber": { "type": "number", "nullable": true },
          "year": { "type": "number", "nullable": true },
          "semester": { "type": "number", "nullable": true },
          "shortCode": { "type": "string" },
          "isOwner": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "author": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PostAuthorEntity" }]
          },
          "course": { "$ref": "#/components/schemas/PostCourseEntity" },
          "files": { "type": "array", "items": { "$ref": "#/components/schemas/PostFileEntity" } },
          "_count": { "$ref": "#/components/schemas/PostCountEntity" },
          "savedByCurrentUser": { "type": "boolean" },
          "views": { "type": "number" },
          "reactionCounts": { "type": "object", "description": "Map of ReactionType to count" },
          "userReaction": { "type": "string", "nullable": true },
          "tags": { "type": "array", "items": { "$ref": "#/components/schemas/PostTagEntity" } }
        },
        "required": [
          "id",
          "type",
          "status",
          "publicationStatus",
          "trendingScore",
          "isAnonymous",
          "shortCode",
          "isOwner",
          "createdAt",
          "updatedAt",
          "course",
          "files",
          "_count",
          "savedByCurrentUser",
          "views",
          "reactionCounts",
          "tags"
        ]
      },
      "PostEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "allOf": [{ "$ref": "#/components/schemas/PostType" }] },
          "status": { "allOf": [{ "$ref": "#/components/schemas/PostStatus" }] },
          "publicationStatus": {
            "allOf": [{ "$ref": "#/components/schemas/PostPublicationStatus" }]
          },
          "trendingScore": { "type": "number" },
          "isAnonymous": { "type": "boolean", "default": false },
          "title": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "externalUrl": { "type": "string", "nullable": true },
          "summary": { "type": "string", "nullable": true },
          "summarizedAt": { "format": "date-time", "type": "string", "nullable": true },
          "contentWarning": { "type": "string", "nullable": true },
          "examYear": { "type": "number", "nullable": true },
          "moduleNumber": { "type": "number", "nullable": true },
          "year": { "type": "number", "nullable": true },
          "semester": { "type": "number", "nullable": true },
          "shortCode": { "type": "string" },
          "isOwner": { "type": "boolean" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "author": {
            "nullable": true,
            "allOf": [{ "$ref": "#/components/schemas/PostAuthorEntity" }]
          },
          "course": { "$ref": "#/components/schemas/PostCourseEntity" },
          "files": { "type": "array", "items": { "$ref": "#/components/schemas/PostFileEntity" } },
          "_count": { "$ref": "#/components/schemas/PostCountEntity" },
          "savedByCurrentUser": { "type": "boolean" },
          "views": { "type": "number" },
          "reactionCounts": { "type": "object", "description": "Map of ReactionType to count" },
          "userReaction": { "type": "string", "nullable": true },
          "tags": { "type": "array", "items": { "$ref": "#/components/schemas/PostTagEntity" } }
        },
        "required": [
          "id",
          "type",
          "status",
          "publicationStatus",
          "trendingScore",
          "isAnonymous",
          "shortCode",
          "isOwner",
          "createdAt",
          "updatedAt",
          "course",
          "files",
          "_count",
          "savedByCurrentUser",
          "views",
          "reactionCounts",
          "tags"
        ]
      },
      "PaginatedPostEntity": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/PostEntity" } },
          "total": { "type": "number" },
          "page": { "type": "number" },
          "limit": { "type": "number" },
          "totalPages": { "type": "number" }
        },
        "required": ["items", "total", "page", "limit", "totalPages"]
      },
      "ReactToPostDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["HELPFUL", "LOVE", "FIRE", "WOW", "SALUTE", "FUNNY"]
          }
        },
        "required": ["type"]
      },
      "UpdateablePostStatus": { "type": "string", "enum": ["APPROVED", "REJECTED"] },
      "UpdatePostStatusDto": {
        "type": "object",
        "properties": {
          "status": { "allOf": [{ "$ref": "#/components/schemas/UpdateablePostStatus" }] }
        },
        "required": ["status"]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 3, "maxLength": 200 },
          "description": { "type": "string", "maxLength": 2000 },
          "isAnonymous": { "type": "boolean", "default": false },
          "externalUrl": { "type": "string", "maxLength": 500 },
          "examYear": { "type": "number", "minimum": 1900, "maximum": 2100 },
          "moduleNumber": { "type": "number", "minimum": 1, "maximum": 20 },
          "year": { "type": "number", "minimum": 1, "maximum": 6 },
          "semester": { "type": "number", "minimum": 1, "maximum": 3 },
          "tags": {
            "example": ["Linear Algebra", "Calculus"],
            "description": "Array of tag names (max 5 tags)",
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "TagPostDto": {
        "type": "object",
        "properties": { "tags": { "type": "array", "items": { "type": "string" } } },
        "required": ["tags"]
      },
...
      "ReadingListEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "isPublic": { "type": "boolean" },
          "postCount": { "type": "number" },
          "postIds": { "type": "array", "items": { "type": "string" } },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" }
        },
        "required": [
          "id",
          "userId",
          "name",
          "isPublic",
          "postCount",
          "postIds",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}